markup2pdf-backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
reportlab
weasyprint
markdown-it-py
mdit-py-plugins
pygments
python-dotenv
